extract correlation lookup route
